video-mgmt/requirements.txt
celery
redis
Flask
Werkzeug
moviepy
Pillow==9.5.0
ffmpeg-python
ffmpeg
opencv-python
opencv-python-headless
psycopg2-binary